requirements.txt
streamlit==0.79.0
pandas==1.2.3
numpy==1.20.1
Pillow==6.2.1
xlsxwriter==1.2.1
nltk==3.5
sklearn==0.23.2